+++ app/locales/ja/translations.ts
@@ -411,6 +411,11 @@ export default {
     sign_up_policy: {
         paragraph: '「無料で登録」をクリックすると、<a href="{{link1}}">利用規約</a>に同意し、<a href="{{link3}}">Cookieの使用</a>に関する情報を含む<a href="{{link2}}">プライバシーポリシー</a>を読んだことになります。', // TODO: update this based on en translation
     },
+    tos_consent: {
+        paragraph: 'We\'ve updated our <a href="{{link1}}">Terms of Use</a> and <a href="{{link2}}">Privacy Policy</a>. Please read them carefully.',
+        have_read_and_agree: 'I have read and agree to these terms.',
+        continue: 'Continue',
+    },
     validationErrors: {
         description: 'このフィールド',
         inclusion: '{{description}}はリストに含まれていません。',