+++ types/ember-data-factory-guy/index.d.ts
@@ -1,2 +1,8 @@
-export function make(factory: string);
+export function make(factory: string, data?: object);
 export function setupFactoryGuy(hooks: any);
+
+class Mock {
+    returns(object): void;
+}
+
+export function mockFindRecord(model: string): Mock;